Refactored code so that we could have controllers. Took a while to figure out how to write the file paths for the JSON and view EJS files...
