+++ dinosaurs.json
@@ -1,1 +1,1 @@
-[{"name":"Littlefoot","type":"apatosaurus"},{"name":"Cera","type":"triceratops"},{"name":"Ducky","type":"saurolophus"},{"name":"Petrie","type":"pteranodon"},{"name":"Spike","type":"stegosaurus"}]
+[{"name":"Littlefoot","type":"apatosaurus"},{"name":"Cera","type":"triceratops"},{"name":"Ducky","type":"saurolophus"},{"name":"Petrie","type":"pteranodon"},{"name":"Spike","type":"stegosaurus"},{"name":"Trisha","type":"Babysaur"},{"name":"Sara","type":"Saratops"}]
